Added files via upload
Basic , forms for the restaurant , basically if a waiter would to access the the menu , this is what will be presented. It's a small rough draft but really if it had a few methods it can be optional to be used. Feel free , to add anything to it or make suggestions. I'll be still adding more to it as the week progresses and will be keeping updated on the forms if I make any big changes.